# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: 新点 SaaS 造价系统 API
  version: 1.0.0-current
  summary: 当前代码已注册的 V1 API 最小契约
  description: |
    自动生成文件，请勿手工编辑。
    生成命令：npm run docs:openapi-current
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ObjectEnvelope:
      type: object
      additionalProperties: true
    ListEnvelope:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
paths:
...
  /v1/ai/bill-recommendations:
    post:
      operationId: postAiBillRecommendations
      summary: POST /v1/ai/bill-recommendations
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/quota-recommendations:
    post:
      operationId: postAiQuotaRecommendations
      summary: POST /v1/ai/quota-recommendations
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/recommendations/{recommendationId}/accept:
    post:
      operationId: postAiRecommendationsByRecommendationIdAccept
      summary: POST /v1/ai/recommendations/:recommendationId/accept
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: recommendationId
        in: path
        description: recommendationId path parameter.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/recommendations/{recommendationId}/expire:
    post:
      operationId: postAiRecommendationsByRecommendationIdExpire
      summary: POST /v1/ai/recommendations/:recommendationId/expire
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: recommendationId
        in: path
        description: recommendationId path parameter.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/recommendations/{recommendationId}/ignore:
    post:
      operationId: postAiRecommendationsByRecommendationIdIgnore
      summary: POST /v1/ai/recommendations/:recommendationId/ignore
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: recommendationId
        in: path
        description: recommendationId path parameter.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/recommendations/expire-stale:
    post:
      operationId: postAiRecommendationsExpireStale
      summary: POST /v1/ai/recommendations/expire-stale
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/variance-warnings:
    post:
      operationId: postAiVarianceWarnings
      summary: POST /v1/ai/variance-warnings
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/ai/bill-recommendations:
    get:
      operationId: getProjectsByProjectIdAiBillRecommendations
      summary: GET /v1/projects/:projectId/ai/bill-recommendations
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: resourceType
        in: query
        description: resourceType query parameter.
        required: false
        schema:
          type: string
      - name: resourceId
        in: query
        description: resourceId query parameter.
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Status filter.
        required: false
        schema:
          type: string
          enum: [queued, processing, completed, failed, pending, approved, rejected, cancelled, draft, submitted, generated, accepted, ignored, expired]
      - name: stageCode
        in: query
        description: Project stage code.
        required: false
        schema:
          type: string
      - name: disciplineCode
        in: query
        description: Project discipline code.
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: Maximum number of records to return.
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/ai/quota-recommendations:
    get:
      operationId: getProjectsByProjectIdAiQuotaRecommendations
      summary: GET /v1/projects/:projectId/ai/quota-recommendations
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: resourceType
        in: query
        description: resourceType query parameter.
        required: false
        schema:
          type: string
      - name: resourceId
        in: query
        description: resourceId query parameter.
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Status filter.
        required: false
        schema:
          type: string
          enum: [queued, processing, completed, failed, pending, approved, rejected, cancelled, draft, submitted, generated, accepted, ignored, expired]
      - name: stageCode
        in: query
        description: Project stage code.
        required: false
        schema:
          type: string
      - name: disciplineCode
        in: query
        description: Project discipline code.
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: Maximum number of records to return.
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/ai/recommendation-context:
    get:
      operationId: getProjectsByProjectIdAiRecommendationContext
      summary: GET /v1/projects/:projectId/ai/recommendation-context
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/ai/recommendations:
    get:
      operationId: getProjectsByProjectIdAiRecommendations
      summary: GET /v1/projects/:projectId/ai/recommendations
      tags:
        - AI Recommendations
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: recommendationType
        in: query
        description: AI recommendation type.
        required: false
        schema:
          type: string
          enum: [bill_recommendation, quota_recommendation, variance_warning]
      - name: resourceType
        in: query
        description: resourceType query parameter.
        required: false
        schema:
          type: string
      - name: resourceId
        in: query
        description: resourceId query parameter.
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Status filter.
        required: false
        schema:
          type: string
          enum: [queued, processing, completed, failed, pending, approved, rejected, cancelled, draft, submitted, generated, accepted, ignored, expired]
      - name: stageCode
        in: query
        description: Project stage code.
        required: false
        schema:
          type: string
      - name: disciplineCode
        in: query
        description: Project discipline code.
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: Maximum number of records to return.
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/ai/variance-warning-thresholds:
    get:
      operationId: getProjectsByProjectIdAiVarianceWarningThresholds
      summary: GET /v1/projects/:projectId/ai/variance-warning-thresholds
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putProjectsByProjectIdAiVarianceWarningThresholds
      summary: PUT /v1/projects/:projectId/ai/variance-warning-thresholds
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...